클컴백엔드/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
apscheduler
alembic
python-multipart
httpx
bcrypt==4.3.0